removes Exceptions as it doesn't neeed now and makes the code more simple

+++ src/main/java/com/college/student/controller/AdmissionController.java
@@ -8,17 +8,17 @@ import com.college.student.service.StudentService;
 import java.util.List;
 import java.util.Scanner;
 public class AdmissionController {
-    private static Scanner scanner = new Scanner(System.in);
-    private static int studentRollNo;
-    private static String studentName;
-    private static byte studentAge;
-    private static long studentPhoneNo;
     public static void main(String[] args) {
-        int storageType = storageTypeOption();
+        Scanner scanner = new Scanner(System.in);
+        int studentRollNo;
+        String studentName;
+        byte studentAge;
+        long studentPhoneNo;
+        String storageType = storageTypeOption(scanner);
         StudentService studentService = new StudentService(storageType);
         int choice = Integer.MAX_VALUE;
         while (true) {
-            choice = displayStudentChoices();
+            choice = displayStudentChoices(scanner,choice);
             switch (choice) {
                 case 1 :
                     Student student = new Student();
@@ -53,42 +53,59 @@ public class AdmissionController {
                     System.out.print("\nEnter Student Roll No to delete from Student List : ");
                     studentRollNo = scanner.nextInt();
                     scanner.nextLine();
-                    Student deletedStudent = studentService.deleteStudentByRollNo(studentRollNo);
-                    System.out.println("below Student Detail : ");
-                    System.out.println(" RollNo |  Name  |  Age  |  PhoneNo");
-                    System.out.println(deletedStudent);
-                    System.out.println("--------------Has Been deleted Successfully From List----------------");
+                    if(isStudentExist(studentService,studentRollNo)) {
+                        Student deletedStudent = studentService.deleteStudentByRollNo(studentRollNo);
+                        System.out.println("below Student Detail : ");
+                        System.out.println(" RollNo |  Name  |  Age  |  PhoneNo");
+                        System.out.println(deletedStudent);
+                        System.out.println("--------------Has Been deleted Successfully From List----------------");
+                    } else {
+                        System.out.println("Student rollNo doesn't Exits !  Please Enter the Correct RollNo !!!");
+                        continue;
+                    }
+
                     break;
                 case 4 :
+                    student = new Student();    //created a new student object
                     System.out.print("\nEnter student RollNo to Update From List : ");
-                    int studentRollNoToUpdate = scanner.nextInt();
-                    scanner.nextLine();
-                    System.out.print("\nEnter new Student RollNo to Update : ");
                     studentRollNo = scanner.nextInt();
                     scanner.nextLine();
-                    System.out.print("\nEnter new Student Name to Update : ");
-                    studentName = scanner.nextLine();
-                    System.out.print("\nEnter new Student Age to Update : ");
-                    studentAge = scanner.nextByte();
-                    scanner.nextLine();
-                    System.out.print("\nEnter new Student phoneNumber to Update : ");
-                    studentPhoneNo = scanner.nextLong();
-                    scanner.nextLine();
-                    Student updatedStudent = studentService.updateStudentDetailsByRollNo(studentRollNoToUpdate, studentRollNo,studentName,studentAge,studentPhoneNo);
-                    System.out.println("Below Student detail :");
-                    System.out.println(" RollNo |  Name  |  Age  |  PhoneNo");
-                    System.out.println(updatedStudent);
-                    System.out.println("-----------------Has Been Updated Successfully----------------");
+                    if(isStudentExist(studentService,studentRollNo)) {
+                        System.out.print("\nEnter new  Name to Update : ");
+                        studentName = scanner.nextLine();
+                        System.out.print("\nEnter new Age to Update : ");
+                        studentAge = scanner.nextByte();
+                        scanner.nextLine();
+                        System.out.print("\nEnter new  phoneNumber to Update : ");
+                        studentPhoneNo = scanner.nextLong();
+                        scanner.nextLine();
+                        student.setRollNo(studentRollNo);
+                        student.setName(studentName);
+                        student.setAge(studentAge);
+                        student.setPhoneNo(studentPhoneNo);
+                        Student updatedStudent = studentService.updateStudentDetailsByRollNo(student);
+                        System.out.println("Below Student detail :");
+                        System.out.println(" RollNo |  Name  |  Age  |  PhoneNo");
+                        System.out.println(updatedStudent);
+                        System.out.println("-----------------Has Been Updated Successfully----------------");
+                    } else {
+                        System.out.println("Student RollNo not Found, Please Enter a Valid RollNo !!!");
+                    }
+
                     break;
                 case 5 :
                     System.out.print("Enter rollNo to get the Student Data : ");
                     studentRollNo = scanner.nextInt();
                     scanner.nextLine();
-                    Student studentDetail = studentService.getStudentByRollNo(studentRollNo);
-                    System.out.println("-------------Student--------------");
-                    System.out.println(" RollNo |  Name  |  Age  |  PhoneNo");
-                    System.out.println(studentDetail);
-                    System.out.println("-------------Student--------------");
+                    if(isStudentExist(studentService,studentRollNo)) {
+                        Student studentDetail = studentService.getStudentByRollNo(studentRollNo);
+                        System.out.println("-------------Student--------------");
+                        System.out.println(" RollNo |  Name  |  Age  |  PhoneNo");
+                        System.out.println(studentDetail);
+                        System.out.println("-------------Student--------------");
+                    } else {
+                        System.out.println("Student RollNo not Found, Please Enter a Valid RollNo !!!");
+                    }
                     break;
                 case 0 :
                     System.exit(0);
@@ -97,23 +114,31 @@ public class AdmissionController {
             }
         }
     }
-    public static int storageTypeOption() {
+    public static String storageTypeOption(Scanner scanner) {
         System.out.println("****************Welcome to College Administration***************");
         System.out.println("where do you want to save the data ?");
-        System.out.println("1. InMemory");
-        System.out.println("2. File Memory");
+        System.out.println("1. InMemory  || inmemory");
+        System.out.println("2. FileMemory || filememory");
         System.out.print("Enter Option : ");
-        int storageType = scanner.nextInt();
-        scanner.nextLine();
+        String storageType = scanner.nextLine();
         return storageType;
     }
-    public static int displayStudentChoices() {
+    public static int displayStudentChoices(Scanner scanner,int choice) {
         System.out.println("1. Add Student In List\n2. Display All Students data");
         System.out.println("3.delete specific student data\n4.update specific student details" );
         System.out.println("5. Get specific student data by rollNo\n0. '0' for EXIT");
         System.out.print("Enter Option : ");
-        int choice = scanner.nextInt();
+        choice = scanner.nextInt();
         scanner.nextLine();
         return choice;
+    }
+    public static boolean isStudentExist(StudentService studentService, int rollNo) {
+        List<Student> studentList = studentService.listStudents();
+        for( Student student : studentList) {   //returns true if student rollNo exists else false;
+            if(student.getRollNo() == rollNo) {
+                return true;
+            }
+        }
+        return false;
     }
 }
